+++ src/agdb/collections/dictionary/dictionary_data_memory.rs
@@ -3,7 +3,7 @@ use super::dictionary_index::DictionaryIndex;
 use super::dictionary_value::DictionaryValue;
 use crate::collections::multi_map::MultiMap;
 use crate::db::db_error::DbError;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 
 pub struct DictionaryDataMemory<T>
