+++ src/agdb/collections/map_common/map_data_storage.rs
@@ -4,7 +4,7 @@ use super::map_value_state::MapValueState;
 use crate::db::db_error::DbError;
 use crate::storage::storage_index::StorageIndex;
 use crate::storage::OldStorage;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
 use std::hash::Hash;
