+++ src/agdb/collections/multi_map/multi_map_impl.rs
@@ -3,7 +3,7 @@ use crate::collections::map_common::map_data::MapData;
 use crate::collections::map_common::map_iterator::MapIterator;
 use crate::collections::map_common::map_value_state::MapValueState;
 use crate::db::db_error::DbError;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use crate::utilities::stable_hash::StableHash;
 use std::hash::Hash;
