+++ src/agdb/graph/graph_data_storage_indexes.rs
@@ -1,6 +1,6 @@
 use crate::db::db_error::DbError;
 use crate::storage::storage_index::StorageIndex;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use std::mem::size_of;
 
 pub(crate) struct GraphDataStorageIndexes {
