+++ src/agdb/storage.rs
@@ -12,8 +12,8 @@ mod write_ahead_log;
 use self::partial_storage::PartialStorage;
 use crate::db::db_error::DbError;
 use crate::storage::storage_index::StorageIndex;
-use crate::utilities::serialize::OldSerialize;
-use crate::utilities::serialize::Serialize;
+use crate::utilities::old_serialize::OldSerialize;
+use crate::utilities::old_serialize::Serialize;
 
 pub trait OldStorage {
     fn commit(&mut self) -> Result<(), DbError>;
