+++ src/agdb/storage/storage_impl.rs
@@ -4,7 +4,7 @@ use crate::storage::storage_index::StorageIndex;
 use crate::storage::storage_record::StorageRecord;
 use crate::storage::write_ahead_log::WriteAheadLogRecord;
 use crate::storage::OldStorage;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use std::cmp::max;
 use std::cmp::min;
 use std::io::SeekFrom;
