+++ src/agdb/storage/write_ahead_log.rs
@@ -1,5 +1,5 @@
 use crate::db::db_error::DbError;
-use crate::utilities::serialize::OldSerialize;
+use crate::utilities::old_serialize::OldSerialize;
 use std::fs::File;
 use std::fs::OpenOptions;
 use std::io::Read;
