+++ src/agdb/utilities.rs
@@ -1,3 +1,3 @@
+pub mod old_serialize;
 pub mod partial_serialize;
-pub mod serialize;
 pub mod stable_hash;
