+++ src/agdb/utilities/old_serialize.rs
